Support reading pkg-config dependencies and versions from Cargo.toml metadata
Add a function `pkg_config::probe_all()` that parses pkg-config package
names and version numbers from a `Cargo.toml` metadata table,
`[package.metadata.pkg-config]`. This makes that metadata declarative,
rather than programmatic, which means other software can parse it as
well.

This does not require any special support in Cargo, since it just uses a
metadata table. And it does not conflict with the use of mechanisms
other than pkg-config to find dependencies on other platforms; this
metadata only applies to the invocation of the pkg-config crate.

Includes tests for both the success and failure cases, with accompanying
directories containing toml files for each test.