No cross compilation if TARGET is missing
Some systems have the $HOST variable set to the short hostname, and
there is no $TARGET variable set there.

Currently we are fetching both variables and defaulting to empty string
is it is missing, and compare both to decide if there will be a cross
compilation.

This patch default $TARGET to the value of $HOST, as a missing $TARGET
can be interpreted of non-intention of any cross compilation.

Signed-off-by: Alberto Planas <aplanas@suse.com>