Preserve library order between file paths and libs specified via flags
pkg-config files can specify library dependencies in two main ways: via
-l or -L flags (e.g. -lstdc++) or via a filesystem paths. Currently,
pkg-config-rs does not preserve the order between them when generating
cargo commands - it prints commands for the first kind, then for the
second. Some linkers are sensitive to the order in which libraries are
specified, so this behavior won't work with pkg-config files in which
a library specified via a path depends on a library specified via an -l
flag.

This commit ensures that the generated cargo commands are correctly
ordered with respect to the order in which libraries are specified
in the pkg-config output.