Never pick up Strawberry Perl's pkg-config as a valid implementation
Strawberry Perl places a `pkg-config.bat` into PATH that is written in
Perl and is not intended to be used by third parties as a MinGW
distribution. This wouldn't matter, except that Strawberry Perl is
also included in Github CI images out of the box, in `PATH`, and it
breaks everyone's CI jobs.

This is already done by Meson and CMake:

https://github.com/mesonbuild/meson/pull/9384

https://gitlab.kitware.com/cmake/cmake/-/merge_requests/9375

Fixes https://github.com/rust-lang/pkg-config-rs/issues/164