Fix static linking against libbsd on ubuntu
I noticed that I was getting a shared dependency on libbsd.so from a
package using pkg-config recently, and it seems to be due to this
changed line; the Ubuntu libbsd pkg-config name is `libbsd`, so
previously `is_static_available` would look for `liblibbsd`.

At a glance, it seems like a few different libraries have `libfoo.pc` as
libbsd does, so I’ve proposed the following change, just adding `{}.a`
to the list after `lib{}.a`.